Implement .env file for setting development environment vars (#23)
* Setup

This is the general idea and how people seem to be doing it in those
articles we read on the weekend. The app seems to hang up though and
I’m not sure where or why. We don’t define a ‘production’ env spec ever
because every tutorial recommends to set up environment variables
manually on your host platform.

* Fix environment variable loading

* remove .env

* combine seeding into db file, ensure connection made before seed

* seed false by default

* PR requests 1